+++ libs/core/package.json
@@ -16,7 +16,10 @@
     "test": "NODE_OPTIONS='--import tsx/esm' NODE_ENV=test mocha 'test/**/*.spec.ts'"
   },
   "dependencies": {
-    "tsx": "^4.7.2",
+    "@hicommonwealth/shared": "*",
     "zod": "^3.22.4"
+  },
+  "devDependencies": {
+    "tsx": "^4.7.2"
   }
 }
