+++ libs/core/src/index.ts
@@ -2,5 +2,3 @@ export * from './errors';
 export * from './framework';
 export * from './ports';
 export * as schemas from './schemas';
-export * from './types';
-export * from './utils';
