+++ libs/core/src/ports/interfaces.ts
@@ -5,12 +5,6 @@ import {
   EventsHandlerMetadata,
 } from '../framework';
 import { Events } from '../schemas';
-import {
-  AnalyticsOptions,
-  BrokerPublications,
-  BrokerSubscriptions,
-  CacheNamespaces,
-} from '../types';
 
 /**
  * Resource disposer function
@@ -50,6 +44,18 @@ export interface Stats extends Disposable {
   gauge(key: string, value: number): void;
   // traces
   timing(key: string, duration: number, tags?: Record<string, string>): void;
+}
+
+export enum CacheNamespaces {
+  Route_Response = 'route_response',
+  Function_Response = 'function_response',
+  Global_Response = 'global_response',
+  Test_Redis = 'test_redis',
+  Database_Cleaner = 'database_cleaner',
+  Compound_Gov_Version = 'compound_gov_version',
+  Token_Balance = 'token_balance',
+  Activity_Cache = 'activity_cache',
+  Rate_Limiter = 'rate_limiter',
 }
 
 /**
@@ -101,6 +107,8 @@ export interface Cache extends Disposable {
   ): Promise<boolean>;
 }
 
+export type AnalyticsOptions = Record<string, any>;
+
 /**
  * Analytics port
  */
@@ -115,6 +123,18 @@ export type RetryStrategyFn = (
   ackOrNackFn: (...args: any[]) => void,
   log: ILogger,
 ) => void;
+
+export enum BrokerPublications {
+  MessageRelayer = 'MessageRelayer',
+  DiscordListener = 'DiscordMessage',
+}
+
+export enum BrokerSubscriptions {
+  SnapshotListener = 'SnapshotListener',
+  DiscordListener = 'DiscordMessage',
+  ChainEvent = 'ChainEvent',
+  NotificationsProvider = 'NotificationsProvider',
+}
 
 /**
  * Broker Port
