+++ libs/core/src/schemas/commands/group.schemas.ts
@@ -1,5 +1,5 @@
+import { MAX_SCHEMA_INT, MIN_SCHEMA_INT } from '@hicommonwealth/shared';
 import { z } from 'zod';
-import { MAX_SCHEMA_INT, MIN_SCHEMA_INT } from '../../constants';
 import { Community, Group, Requirement } from '../entities.schemas';
 import { PG_INT } from '../utils.schemas';
