+++ libs/core/src/schemas/queries/community.schemas.ts
@@ -1,5 +1,5 @@
+import { MAX_SCHEMA_INT, MIN_SCHEMA_INT } from '@hicommonwealth/shared';
 import { z } from 'zod';
-import { MAX_SCHEMA_INT, MIN_SCHEMA_INT } from '../../constants';
 import { CommunityMember, CommunityStake } from '../entities.schemas';
 import { PG_INT } from '../utils.schemas';
 import { PaginatedResultSchema, PaginationParamsSchema } from './pagination';
