+++ libs/model/src/services/tokenBalanceCache/tokenBalanceCache.ts
@@ -1,5 +1,6 @@
-import { BalanceSourceType, stats } from '@hicommonwealth/core';
+import { stats } from '@hicommonwealth/core';
 import { logger } from '@hicommonwealth/logging';
+import { BalanceSourceType } from '@hicommonwealth/shared';
 import { fileURLToPath } from 'url';
 import { getCosmosBalances, getEvmBalances } from './providers';
 import {
