+++ libs/shared/src/constants.ts
@@ -1,3 +1,6 @@
+export const ALL_COMMUNITIES = 'all_communities';
+export const MAX_COMMUNITY_IMAGE_SIZE_KB = 500;
+
 export const MIN_SCHEMA_INT = 0;
 export const MAX_SCHEMA_INT = 2147483647;
