+++ libs/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * as commonProtocol from './commonProtocol';
-export * from './shared-types';
+export * from './constants';
+export * from './types';
 export * from './utils';
