+++ libs/shared/src/types/index.ts
@@ -1,0 +1,5 @@
+export * from './discord';
+export * from './notification';
+export * from './proposal';
+export * from './protocol';
+export * from './types';
