+++ libs/shared/src/types/protocol.ts
@@ -1,0 +1,176 @@
+export enum BalanceSourceType {
+  ETHNative = 'eth_native',
+  ERC20 = 'erc20',
+  ERC721 = 'erc721',
+  ERC1155 = 'erc1155',
+  CosmosNative = 'cosmos_native',
+  CW20 = 'cw20',
+  CW721 = 'cw721',
+}
+
+export enum BalanceType {
+  Terra = 'terra',
+  Ethereum = 'ethereum',
+  Solana = 'solana',
+  Cosmos = 'cosmos',
+  NEAR = 'near',
+  Substrate = 'substrate',
+}
+
+export enum SupportedNetwork {
+  Substrate = 'substrate',
+  Aave = 'aave',
+  Compound = 'compound',
+  ERC20 = 'erc20',
+  ERC721 = 'erc721',
+  ERC1155 = 'erc1155',
+  Cosmos = 'cosmos',
+}
+
+export type ContractSource = {
+  source_type:
+    | BalanceSourceType.ERC20
+    | BalanceSourceType.ERC721
+    | BalanceSourceType.ERC1155;
+  evm_chain_id: number;
+  contract_address: string;
+  token_id?: string;
+};
+
+export type NativeSource = {
+  source_type: BalanceSourceType.ETHNative;
+  evm_chain_id: number;
+};
+
+export type CosmosSource = {
+  source_type: BalanceSourceType.CosmosNative;
+  cosmos_chain_id: string;
+  token_symbol: string;
+};
+
+export type CosmosContractSource = {
+  source_type: BalanceSourceType.CW20 | BalanceSourceType.CW721;
+  cosmos_chain_id: string;
+  contract_address: string;
+};
+
+export type ThresholdData = {
+  threshold: string;
+  source: ContractSource | NativeSource | CosmosSource | CosmosContractSource;
+};
+
+export type AbiType = Record<string, unknown>[];
+
+export enum WalletId {
+  Magic = 'magic',
+  Polkadot = 'polkadot',
+  Metamask = 'metamask',
+  WalletConnect = 'walletconnect',
+  KeplrEthereum = 'keplr-ethereum',
+  Keplr = 'keplr',
+  Leap = 'leap',
+  NearWallet = 'near',
+  TerraStation = 'terrastation',
+  TerraWalletConnect = 'terra-walletconnect',
+  CosmosEvmMetamask = 'cosm-metamask',
+  Phantom = 'phantom',
+  Coinbase = 'coinbase',
+}
+
+// 'google', 'github', 'discord', and 'twitter' are passed to magic login directly
+export enum WalletSsoSource {
+  Google = 'google',
+  Github = 'github',
+  Discord = 'discord',
+  Twitter = 'twitter',
+  apple = 'apple',
+  Email = 'email',
+  Unknown = 'unknown', // address created after we launched SSO, before we started recording WalletSsoSource
+}
+
+export enum ChainBase {
+  CosmosSDK = 'cosmos',
+  Substrate = 'substrate',
+  Ethereum = 'ethereum',
+  NEAR = 'near',
+  Solana = 'solana',
+}
+
+export enum ContractType {
+  AAVE = 'aave',
+  COMPOUND = 'compound',
+  ERC20 = 'erc20',
+  ERC721 = 'erc721',
+  MARLINTESTNET = 'marlin-testnet',
+  SPL = 'spl',
+  COMMONPROTOCOL = 'common-protocol',
+}
+
+export enum ChainType {
+  Chain = 'chain',
+  DAO = 'dao',
+  Token = 'token',
+  Offchain = 'offchain',
+}
+
+export enum CommunityCategoryType {
+  DeFi = 'DeFi',
+  DAO = 'DAO',
+}
+
+// TODO: remove many of these chain networks, esp substrate (make them all "Substrate"),
+// and just use id to identify specific chains for conditionals
+export enum ChainNetwork {
+  Edgeware = 'edgeware',
+  EdgewareTestnet = 'edgeware-testnet',
+  Kusama = 'kusama',
+  Kulupu = 'kulupu',
+  Polkadot = 'polkadot',
+  Plasm = 'plasm',
+  Stafi = 'stafi',
+  Darwinia = 'darwinia',
+  Phala = 'phala',
+  Centrifuge = 'centrifuge',
+  Straightedge = 'straightedge',
+  Osmosis = 'osmosis',
+  Injective = 'injective',
+  InjectiveTestnet = 'injective-testnet',
+  Terra = 'terra',
+  Ethereum = 'ethereum',
+  NEAR = 'near',
+  NEARTestnet = 'near-testnet',
+  Compound = 'compound',
+  Aave = 'aave',
+  AaveLocal = 'aave-local',
+  dYdX = 'dydx',
+  Metacartel = 'metacartel',
+  ALEX = 'alex',
+  ERC20 = 'erc20',
+  ERC721 = 'erc721',
+  ERC1155 = 'erc1155',
+  CW20 = 'cw20',
+  CW721 = 'cw721',
+  Clover = 'clover',
+  HydraDX = 'hydradx',
+  Crust = 'crust',
+  Sputnik = 'sputnik',
+  SolanaDevnet = 'solana-devnet',
+  SolanaTestnet = 'solana-testnet',
+  Solana = 'solana',
+  SPL = 'spl', // solana token
+  Evmos = 'evmos',
+  Kava = 'kava',
+  Kyve = 'kyve',
+  Stargaze = 'stargaze',
+  Cosmos = 'cosmos',
+}
+
+/**
+ * Cosmos gov module version of a chain
+ */
+export enum CosmosGovernanceVersion {
+  v1 = 'v1',
+  v1beta1 = 'v1beta1',
+  v1beta1Failed = 'v1beta1-attempt-failed',
+  v1Failed = 'v1-attempt-failed',
+}
