+++ libs/tsconfig.json
@@ -4,6 +4,7 @@
     "module": "es2020",
     "target": "esnext",
     "baseUrl": ".",
+    "rootDir": ".",
     "composite": true,
     "strict": true,
 
@@ -15,7 +16,12 @@
 
     "sourceMap": true,
     "declaration": true,
-    "declarationMap": true
+    "declarationMap": true,
+
+    "paths": {
+      "@hicommonwealth/*": ["./*/src"]
+    }
   },
+  "include": ["**/src", "**/test"],
   "exclude": ["**/node_modules", "**/build"]
 }
