+++ packages/commonwealth/server/controllers/server_groups_methods/update_group.ts
@@ -1,4 +1,4 @@
-import { AppError, Requirement, schemas } from '@hicommonwealth/core';
+import { AppError, schemas } from '@hicommonwealth/core';
 import {
   AddressInstance,
   GroupAttributes,
@@ -6,6 +6,7 @@ import {
   UserInstance,
   sequelize,
 } from '@hicommonwealth/model';
+import { Requirement } from '@hicommonwealth/shared';
 import { Op } from 'sequelize';
 import z from 'zod';
 import { MixpanelCommunityInteractionEvent } from '../../../shared/analytics/types';
