+++ packages/commonwealth/server/routes/comments/update_comment_handler.ts
@@ -1,5 +1,5 @@
-import { IDiscordMeta } from '@hicommonwealth/core';
 import { CommentAttributes } from '@hicommonwealth/model';
+import { IDiscordMeta } from '@hicommonwealth/shared';
 import { ServerControllers } from '../../routing/router';
 import { TypedRequest, TypedResponse, success } from '../../types';
