+++ packages/commonwealth/server/routes/threads/create_thread_handler.ts
@@ -1,5 +1,5 @@
-import { IDiscordMeta } from '@hicommonwealth/core';
 import { ThreadAttributes } from '@hicommonwealth/model';
+import { IDiscordMeta } from '@hicommonwealth/shared';
 import { verifyThread } from '../../../shared/canvas/serverVerify';
 import { ServerControllers } from '../../routing/router';
 import { TypedRequestBody, TypedResponse, success } from '../../types';
